requirements.txt
Appium-Python-Client==2.2.0
async-generator==1.10
attrs==21.4.0
behave==1.2.7.dev2
certifi==2021.10.8
cffi==1.15.0
colorama==0.4.4
cryptography==37.0.2
cucumber-tag-expressions==4.1.0
h11==0.13.0
idna==3.3
outcome==1.1.0
parse==1.19.0
parse-type==0.6.0
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
selenium==4.1.5
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
wsproto==1.1.0
